Make use of loading state in book
